// src/services/providers/premium-car/types/premium-car-valuation-response.ts
export type PremiumCarValuationResponse = {
    root: {
      RegistrationDate: Date;
      RegistrationYear: number;
      RegistrationMonth: number;
      ValuationPrivateSaleMinimum: number;
      ValuationPrivateSaleMaximum: number;
      ValuationDealershipMinimum: number;
      ValuationDealershipMaximum: number;
    };
};